New branch with UUID cj6qkzi92umbg3jwb5slxhwlh
Initial branch nickname: source_level_linker

This branch shall develop and maintain a set of scripts for
"source-level linking".

That is, they support chopping large source texts into individual
snippets which can then be sorted into different collections for
potential re-use later.

The individual snippets then represent something like "object files" or
"a library archive" when compared to the typical C development cycle.
(Of course, the snippets are really just text.)

Then other scripts allow to re-assemble the original source files from
the snippets, "linking" them together.

The first iteration of the utilities will be simple - snippets as *.part
files and *.order files which specify in which ordert to reassemble the
snippets into files.

There is no dependency tracking or something fancy yet. This may be
added in later iteration of the utilities, which will also be part of
this branch.

The scripts avoid non-portable behavior and stick to POSIX-defined
features only.