Move branch-related stuff to top-level directory
Things that have formerly been namespaced by putting them into
subdirectories related to this branch move to the top-level directory of
the branch now, because this branch will become a git submodule.

Submodules will be checked out in subdirectories of the master project
anyway, so there is no need to put everything into a subdirectory right
here.